+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.1
+
+### Improvements
+
+-   [💎] Renaming `Custom Marker Icon URL` to `Custom marker URL` to follow Obsidian guidelines
+-   [💎] Showing a toast notice if the user uses an invalid combination of maki icon and custom marker url
+-   [💎] Showing a notice if the plugin has updated so that the user may check for new features
+
 ## 1.1.0
 
 ### Features